Crate spin_templates

Package for working with Wasm component templates.

Structs

InstallOptions

Options controlling template installation.

InstallationResults

The results of installing a set of templates.

ListResults

The result of listing templates.

Run

Executes a template to the point where it is ready to generate artefacts.

RunOptions

Options controlling the execution of a template.

Template

A Spin template.

TemplateManager

Provides access to and operations on the set of installed templates.

Enums

InstalledTemplateWarning

A recoverable problem while listing templates.

SkippedReason

The reason a template was skipped during installation.

TemplateSource

A source from which to install templates.

TemplateVariantInfo

The variant mode in which a template should be run.

Traits

ProgressReporter

Used during template installation to report progress and current activity.